requirements.txt
# Core Framework
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0

# Database
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
asyncpg==0.29.0

# Web Scraping
# Pin to 0.25.2 to satisfy python-telegram-bot and openai/httpx constraints during container builds
httpx==0.25.2
beautifulsoup4==4.12.3
playwright==1.41.0
selenium==4.17.2
scrapy==2.11.0
aiohttp==3.9.3

# OCR & Image Processing
pytesseract==0.3.10
Pillow==10.2.0
google-cloud-vision==3.7.0
opencv-python==4.9.0.80

# AI/ML
openai==1.10.0  # For Grok API
scikit-learn==1.4.0
pandas==2.2.0
numpy==1.26.3

# Twitter/X API
tweepy==4.14.0
twitter-api-client==0.10.22

# Social Media
instaloader==4.10.3
praw==7.7.1  # Reddit API

# Blockchain
web3==6.15.0
eth-account==0.11.0

# Messaging / PubSub
redis==5.0.1

# Alerts
discord-webhook==1.3.1
twilio==8.12.0
sendgrid==6.11.0
python-telegram-bot==20.7

# Dashboard
streamlit==1.30.0
plotly==5.18.0
altair==5.2.0

# Scheduling
apscheduler==3.10.4
celery==5.3.6
redis==5.0.1

# Utilities
loguru==0.7.2
tenacity==8.2.3
pyyaml==6.0.1
python-multipart==0.0.6
pytz==2024.1

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
pytest-cov==4.1.0
faker==22.5.1

# Development
black==24.1.1
flake8==7.0.0
mypy==1.8.0
ipython==8.20.0
